Fix WSG FileAccess schema import due to AnyClass relationship

+++ iModelBridgeCore/WSClient/Tests/IntegrationTests/Cache/CachingDataSourceTests.cpp
@@ -78,6 +78,26 @@ TEST_F(CachingDataSourceTests, OpenOrCreate_BentleyConnectGlobal_Succeeds)
     ASSERT_TRUE(objResult.IsSuccess());
     }
 
+TEST_F(CachingDataSourceTests, OpenOrCreate_BentleyConnectGlobalDev_Succeeds)
+    {
+    UrlProvider::Initialize(UrlProvider::Dev, UrlProvider::DefaultTimeout, &m_localState);
+    auto proxy = ProxyHttpHandler::GetFiddlerProxyIfReachable();
+
+    Utf8String serverUrl = "https://dev-wsg20-eus.cloudapp.net/";
+    Utf8String repositoryId = "BentleyCONNECT.Global--CONNECT.GLOBAL";
+    Credentials credentials("8cc45bd041514b58947ea6c09c@gmail.com", "qwe12312");
+    BeFileName cachePath = GetTestCachePath();
+
+    auto manager = ConnectSignInManager::Create(StubValidClientInfo(), proxy, &m_localState);
+    ASSERT_TRUE(manager->SignInWithCredentials(credentials)->GetResult().IsSuccess());
+    auto authHandler = manager->GetAuthenticationHandler(serverUrl, proxy);
+
+    auto client = WSRepositoryClient::Create(serverUrl, repositoryId, StubValidClientInfo(), nullptr, authHandler);
+
+    auto result = CachingDataSource::OpenOrCreate(client, cachePath, StubCacheEnvironemnt())->GetResult();
+    ASSERT_FALSE(nullptr == result.GetValue());
+    }
+
 TEST_F(CachingDataSourceTests, OpenOrCreate_BentleyConnectSharedContent_Succeeds)
     {
     auto proxy = ProxyHttpHandler::GetFiddlerProxyIfReachable();
@@ -101,8 +121,8 @@ TEST_F(CachingDataSourceTests, OpenOrCreate_BentleyConnectSharedContent_Succeeds
     CachedResponseKey key(txn.GetCache().FindOrCreateRoot(nullptr), "Foo");
     txn.Commit();
 
-    auto objResult = ds->GetObjects(key, WSQuery("SharedContentSchema", "Project"), 
-        ICachingDataSource::DataOrigin::RemoteOrCachedData, nullptr, nullptr)->GetResult();
+    auto objResult = ds->GetObjects(key, WSQuery("SharedContentSchema", "Project"),
+                                    ICachingDataSource::DataOrigin::RemoteOrCachedData, nullptr, nullptr)->GetResult();
     ASSERT_TRUE(objResult.IsSuccess());
     }
 
@@ -274,8 +294,8 @@ TEST_F(CachingDataSourceTests, OpenOrCreate_WSG2xProjectWisePluginMapMobileRepos
     Credentials creds("twc", "admin");
     BeFileName cachePath = GetTestCachePath();
 
-    auto info = std::make_shared<ClientInfo>("Bentley-MapMobile", BeVersion(5,4), "77def89a-7e50-4f0e-a4c7-24fb6044dbfb", 
-        "CLQIqB7y8eCUpdJe5uyRVVaaGbk=", "Windows 6.1", nullptr);
+    auto info = std::make_shared<ClientInfo>("Bentley-MapMobile", BeVersion(5, 4), "77def89a-7e50-4f0e-a4c7-24fb6044dbfb",
+                                             "CLQIqB7y8eCUpdJe5uyRVVaaGbk=", "Windows 6.1", nullptr);
 
     IWSRepositoryClientPtr client = WSRepositoryClient::Create(serverUrl, repositoryId, StubValidClientInfo(), nullptr, proxy);
     client->SetCredentials(creds);
@@ -304,33 +324,35 @@ TEST_F(CachingDataSourceTests, SyncLocalChanges_WSG23ProjectWisePluginRepository
     ObjectId uploadsFolderId {"PW_WSG", "Project", "30ff82ee-0a91-42ca-9dd6-c941e577be94"};
     ObjectId uploadsFolderNavNodeId {"Navigation", "NavNode", "ECObjects--" + uploadsFolderId.schemaName + "-" + uploadsFolderId.className + "-" + uploadsFolderId.remoteId};
 
-    {
-    auto txn = ds->StartCacheTransaction();
-    ASSERT_EQ(SUCCESS, txn.GetCache().LinkInstanceToRoot(nullptr, uploadsFolderId));
-    ASSERT_EQ(SUCCESS, txn.GetCache().LinkInstanceToRoot(nullptr, uploadsFolderNavNodeId));
-    txn.Commit();
-    }
+    if (true)
+        {
+        auto txn = ds->StartCacheTransaction();
+        ASSERT_EQ(SUCCESS, txn.GetCache().LinkInstanceToRoot(nullptr, uploadsFolderId));
+        ASSERT_EQ(SUCCESS, txn.GetCache().LinkInstanceToRoot(nullptr, uploadsFolderNavNodeId));
+        txn.Commit();
+        }
 
     auto navigationResult = ds->GetNavigationChildren(uploadsFolderNavNodeId, CachingDataSource::DataOrigin::RemoteData, nullptr)->GetResult();
     ASSERT_TRUE(navigationResult.IsSuccess());
 
-    {
-    auto txn = ds->StartCacheTransaction();
-    auto& uploadedInstanceNavNodes = navigationResult.GetValue().GetJson();
-    for (auto& uploadedInstanceNavNode : uploadedInstanceNavNodes)
+    if (true)
         {
-        ObjectId uploadedInstanceId
-            (
+        auto txn = ds->StartCacheTransaction();
+        auto& uploadedInstanceNavNodes = navigationResult.GetValue().GetJson();
+        for (auto& uploadedInstanceNavNode : uploadedInstanceNavNodes)
+            {
+            ObjectId uploadedInstanceId
+                (
                 uploadedInstanceNavNode["Key_SchemaName"].asString(),
                 uploadedInstanceNavNode["Key_ClassName"].asString(),
                 uploadedInstanceNavNode["Key_InstanceId"].asString()
                 );
 
-        ASSERT_EQ(SUCCESS, txn.GetCache().LinkInstanceToRoot("DeletedInstances", uploadedInstanceId));
-        ASSERT_EQ(SUCCESS, txn.GetCache().GetChangeManager().DeleteObject(txn.GetCache().FindInstance(uploadedInstanceId)));
+            ASSERT_EQ(SUCCESS, txn.GetCache().LinkInstanceToRoot("DeletedInstances", uploadedInstanceId));
+            ASSERT_EQ(SUCCESS, txn.GetCache().GetChangeManager().DeleteObject(txn.GetCache().FindInstance(uploadedInstanceId)));
+            }
+        txn.Commit();
         }
-    txn.Commit();
-    }
 
     Json::Value newDocument = ToJson(
         R"({
@@ -338,21 +360,22 @@ TEST_F(CachingDataSourceTests, SyncLocalChanges_WSG23ProjectWisePluginRepository
         "CreateTime" : "2014-06-03T08:59:58.673Z"
         })");
 
-    {
-    auto txn = ds->StartCacheTransaction();
-    auto ecClass = txn.GetCache().GetAdapter().GetECClass("PW_WSG", "Document");
-    auto ecRelClass = txn.GetCache().GetAdapter().GetECRelationshipClass("PW_WSG", "DocumentParent");
+    if (true)
+        {
+        auto txn = ds->StartCacheTransaction();
+        auto ecClass = txn.GetCache().GetAdapter().GetECClass("PW_WSG", "Document");
+        auto ecRelClass = txn.GetCache().GetAdapter().GetECRelationshipClass("PW_WSG", "DocumentParent");
 
-    auto documentKey = txn.GetCache().GetChangeManager().CreateObject(*ecClass, newDocument);
+        auto documentKey = txn.GetCache().GetChangeManager().CreateObject(*ecClass, newDocument);
 
-    ASSERT_EQ(SUCCESS, txn.GetCache().GetChangeManager().ModifyFile(documentKey, StubFile("TestContent", "'Foo file'.txt"), false));
-    ASSERT_EQ(SUCCESS, txn.GetCache().LinkInstanceToRoot("NewDocument", txn.GetCache().FindInstance(documentKey)));
+        ASSERT_EQ(SUCCESS, txn.GetCache().GetChangeManager().ModifyFile(documentKey, StubFile("TestContent", "'Foo file'.txt"), false));
+        ASSERT_EQ(SUCCESS, txn.GetCache().LinkInstanceToRoot("NewDocument", txn.GetCache().FindInstance(documentKey)));
 
-    auto folderKey = txn.GetCache().FindInstance(uploadsFolderId);
-    auto relationshipKey = txn.GetCache().GetChangeManager().CreateRelationship(*ecRelClass, documentKey, folderKey);
-    ASSERT_TRUE(relationshipKey.IsValid());
-    txn.Commit();
-    }
+        auto folderKey = txn.GetCache().FindInstance(uploadsFolderId);
+        auto relationshipKey = txn.GetCache().GetChangeManager().CreateRelationship(*ecRelClass, documentKey, folderKey);
+        ASSERT_TRUE(relationshipKey.IsValid());
+        txn.Commit();
+        }
 
     Json::Value newProject = ToJson(
         R"({
@@ -360,36 +383,38 @@ TEST_F(CachingDataSourceTests, SyncLocalChanges_WSG23ProjectWisePluginRepository
         "CreateTime" : "2014-06-03T08:59:58.673Z"
         })");
 
-    {
-    auto txn = ds->StartCacheTransaction();
-    auto ecClass = txn.GetCache().GetAdapter().GetECClass("PW_WSG", "Project");
-    auto ecRelClass = txn.GetCache().GetAdapter().GetECRelationshipClass("PW_WSG", "ProjectParent");
+    if (true)
+        {
+        auto txn = ds->StartCacheTransaction();
+        auto ecClass = txn.GetCache().GetAdapter().GetECClass("PW_WSG", "Project");
+        auto ecRelClass = txn.GetCache().GetAdapter().GetECRelationshipClass("PW_WSG", "ProjectParent");
 
-    auto projectKey = txn.GetCache().GetChangeManager().CreateObject(*ecClass, newProject);
-    ASSERT_EQ(SUCCESS, txn.GetCache().LinkInstanceToRoot("NewProject", txn.GetCache().FindInstance(projectKey)));
-    auto folderKey = txn.GetCache().FindInstance(uploadsFolderId);
-    auto relationshipKey = txn.GetCache().GetChangeManager().CreateRelationship(*ecRelClass, projectKey, folderKey);
-    ASSERT_TRUE(relationshipKey.IsValid());
-    txn.Commit();
-    }
+        auto projectKey = txn.GetCache().GetChangeManager().CreateObject(*ecClass, newProject);
+        ASSERT_EQ(SUCCESS, txn.GetCache().LinkInstanceToRoot("NewProject", txn.GetCache().FindInstance(projectKey)));
+        auto folderKey = txn.GetCache().FindInstance(uploadsFolderId);
+        auto relationshipKey = txn.GetCache().GetChangeManager().CreateRelationship(*ecRelClass, projectKey, folderKey);
+        ASSERT_TRUE(relationshipKey.IsValid());
+        txn.Commit();
+        }
 
     auto syncResult = ds->SyncLocalChanges(nullptr, nullptr)->GetResult();
     ASSERT_TRUE(syncResult.IsSuccess());
 
-    {
-    auto txn = ds->StartCacheTransaction();
-    Json::Value newInstances;
-    status = txn.GetCache().ReadInstancesLinkedToRoot("NewDocument", newInstances);
-    ASSERT_EQ(SUCCESS, status);
-    ASSERT_EQ(1, newInstances.size());
-    auto newInstance = newInstances[0];
+    if (true)
+        {
+        auto txn = ds->StartCacheTransaction();
+        Json::Value newInstances;
+        status = txn.GetCache().ReadInstancesLinkedToRoot("NewDocument", newInstances);
+        ASSERT_EQ(SUCCESS, status);
+        ASSERT_EQ(1, newInstances.size());
+        auto newInstance = newInstances[0];
 
-    auto newKey = txn.GetCache().GetAdapter().GetInstanceKeyFromJsonInstance(newInstance);
+        auto newKey = txn.GetCache().GetAdapter().GetInstanceKeyFromJsonInstance(newInstance);
 
-    status = txn.GetCache().GetChangeManager().ModifyFile(newKey, StubFile("TestContentModified"), false);
-    ASSERT_EQ(SUCCESS, status);
-    txn.Commit();
-    }
+        status = txn.GetCache().GetChangeManager().ModifyFile(newKey, StubFile("TestContentModified"), false);
+        ASSERT_EQ(SUCCESS, status);
+        txn.Commit();
+        }
 
     syncResult = ds->SyncLocalChanges(nullptr, nullptr)->GetResult();
     ASSERT_TRUE(syncResult.IsSuccess());
@@ -458,12 +483,12 @@ TEST_F(CachingDataSourceTests, GetObjects_WSG2PWSpatialQuery_Succeeds)
 
     query->SetSelect("*,SpatialLocation.*");
     query->SetFilter("geo.intersects("
-        "SpatialLocation.Location,geometry'Polygon(("
-        "20.896001+53.065279,"
-        "20.896001+57.105292,"
-        "26.872565+57.105292,"
-        "26.872565+53.065279,"
-        "20.896001+53.065279))')");
+                     "SpatialLocation.Location,geometry'Polygon(("
+                     "20.896001+53.065279,"
+                     "20.896001+57.105292,"
+                     "26.872565+57.105292,"
+                     "26.872565+53.065279,"
+                     "20.896001+53.065279))')");
     query->SetSkip(0);
     query->SetTop(100);
 
@@ -651,7 +676,7 @@ TEST_F(CachingDataSourceTests, GetObjects_PunchlistQueries_Succeeds)
             auto relClass = txn.GetCache().GetAdapter().GetECRelationshipClass("Issue.IssueAttachmentRel");
             ECInstanceKey result(resultClass->GetId(), txn.GetCache().GetAdapter().FindInstance(resultClass));
             Json::Value properties;
-            properties["FileName"] = Utf8PrintfString("Test-%d-%s.txt", i, BeGuid().ToString().c_str()) ;
+            properties["FileName"] = Utf8PrintfString("Test-%d-%s.txt", i, BeGuid().ToString().c_str());
             auto attachment = txn.GetCache().GetChangeManager().CreateObject(*attachmentClass, properties);
             auto relationship = txn.GetCache().GetChangeManager().CreateRelationship(*relClass, result, attachment);
             ASSERT_TRUE(relationship.IsValid());
